+++ nextjs-blog-starter-mdx/src/design/components/README.md
@@ -1,0 +1,1 @@
+These are all shared components for all layouts and whatever
