+++ nextjs-blog-starter-mdx/src/design/components/avatar.js
