+++ nextjs-blog-starter-mdx/src/design/components/cover-image.js
