+++ nextjs-blog-starter-mdx/src/design/layout/footer.js
@@ -1,5 +1,4 @@
-import Container from './container'
-import { EXAMPLE_PATH } from '../lib/constants'
+import Container from '../components/container'
 
 export default function Footer() {
   return (
@@ -17,7 +16,7 @@ export default function Footer() {
               Read Documentation
             </a>
             <a
-              href={`https://github.com/vercel/next.js/tree/canary/examples/${EXAMPLE_PATH}`}
+              href={'https://github.com/vercel/next.js/tree/canary/examples/blog-starter'}
               className="mx-3 font-bold hover:underline"
             >
               View on GitHub
