+++ nextjs-blog-starter-mdx/src/design/layout/index.js
@@ -1,6 +1,6 @@
 import Alert from '../components/alert'
-import Footer from '../components/footer'
-import Meta from '../components/meta'
+import Footer from './footer'
+import Meta from './meta'
 
 export default function Layout({ preview, children }) {
   return (
