+++ nextjs-blog-starter-mdx/src/design/layout/meta.js
@@ -1,5 +1,6 @@
 import Head from 'next/head'
-import { CMS_NAME, HOME_OG_IMAGE_URL } from '../lib/constants'
+
+const HOME_OG_IMAGE_URL = 'https://og-image.now.sh/Next.js%20Blog%20Starter%20Example.png?theme=light&md=1&fontSize=100px&images=https%3A%2F%2Fassets.vercel.com%2Fimage%2Fupload%2Ffront%2Fassets%2Fdesign%2Fnextjs-black-logo.svg'
 
 export default function Meta() {
   return (
@@ -34,7 +35,7 @@ export default function Meta() {
       <link rel="alternate" type="application/rss+xml" href="/feed.xml" />
       <meta
         name="description"
-        content={`A statically generated blog example using Next.js and ${CMS_NAME}.`}
+        content={'A statically generated blog example using Next.js and Markdown.'}
       />
       <meta property="og:image" content={HOME_OG_IMAGE_URL} />
     </Head>
