+++ nextjs-blog-starter-mdx/src/design/pages/README.md
@@ -1,0 +1,1 @@
+Page specific components. Not to be shared elsewhere.
