+++ nextjs-blog-starter-mdx/src/design/pages/intro.js
@@ -1,5 +1,3 @@
-import { CMS_NAME } from '../lib/constants'
-
 export default function Intro() {
   return (
     <section className="flex-col md:flex-row flex items-center md:justify-between mt-16 mb-16 md:mb-12">
@@ -14,7 +12,7 @@ export default function Intro() {
         >
           Next.js
         </a>{' '}
-        and {CMS_NAME}.
+        and Markdown.
       </h4>
     </section>
   )
