+++ nextjs-blog-starter-mdx/src/design/post/README.md
@@ -1,0 +1,1 @@
+These are post specific components. Move under `../pages`
