+++ nextjs-blog-starter-mdx/src/design/post/post-body.js
@@ -1,4 +1,4 @@
-import markdownStyles from './markdown-styles.module.css'
+import markdownStyles from '../markdown-styles.module.css'
 
 export default function PostBody({ children }) {
   return (
