--- nextjs-blog-starter-mdx/src/design/section-separator.js
@@ -1,3 +1,0 @@
-export default function SectionSeparator() {
-  return <hr className="border-accent-2 mt-28 mb-24" />
-}
