--- nextjs-blog-starter-mdx/src/lib/constants.js
@@ -1,3 +1,0 @@
-export const EXAMPLE_PATH = 'blog-starter'
-export const CMS_NAME = 'Markdown'
-export const HOME_OG_IMAGE_URL = 'https://og-image.now.sh/Next.js%20Blog%20Starter%20Example.png?theme=light&md=1&fontSize=100px&images=https%3A%2F%2Fassets.vercel.com%2Fimage%2Fupload%2Ffront%2Fassets%2Fdesign%2Fnextjs-black-logo.svg'
