+++ nextjs-blog-starter-mdx/src/lib/markdown/README.md
@@ -1,0 +1,3 @@
+# hey there friendo
+
+This will need to become it's own package.
