+++ nextjs-blog-starter-mdx/src/pages/posts/[...slug].js
@@ -1,16 +1,15 @@
 import { useRouter } from 'next/router'
 import ErrorPage from 'next/error'
-import Container from '../../components/container'
-import PostBody from '../../components/post-body'
-import Header from '../../components/header'
-import PostHeader from '../../components/post-header'
-import Layout from '../../components/layout'
-import PostTitle from '../../components/post-title'
+import Container from '../../design/components/container'
+import PostBody from '../../design/post/post-body'
+import Header from '../../design/layout/header'
+import PostHeader from '../../design/post/post-header'
+import Layout from '../../design/layout'
+import PostTitle from '../../design/post/post-title'
 import Head from 'next/head'
-import { CMS_NAME } from '../../lib/constants'
-import { mdxToString, stringToMdx} from '../../lib/mdxSerialization'
+import { mdxToString, stringToMdx} from '../../lib/markdown/mdxSerialization'
 
-import { getAllPosts } from '../../lib/api'
+import { getAllPosts } from '../../lib/markdown/api'
 
 export default function Post({ post, morePosts, preview }) {
   const router = useRouter()
@@ -36,7 +35,7 @@ export default function Post({ post, morePosts, preview }) {
             <article className="mb-32">
               <Head>
                 <title>
-                  {post.title} | Next.js Blog Example with {CMS_NAME}
+                  {post.title} | Next.js Blog Example with Markdown
                 </title>
               </Head>
               <h1>Title: {post.title}</h1>
